feat(auth): implement JWT authentication with Passport strategy and guards feat(posts): add protected Posts module linked to authenticated users docs(swagger): add API documentation for auth and posts endpoints

+++ prisma/seed.ts
@@ -1,44 +1,56 @@
 import { PrismaClient } from '@prisma/client';
-import { adapter, pool } from '../src/infra/prisma/prisma.setup';
-
-const prisma = new PrismaClient({ adapter });
+import { createPrismaAdapter } from '../src/infra/prisma/prisma.setup';
+import 'dotenv/config';
+import * as bcrypt from 'bcrypt';
 
 async function main() {
-  await prisma.post.deleteMany();
-  await prisma.user.deleteMany();
+  const dbUrl = process.env.DATABASE_URL;
+  if (!dbUrl) throw new Error('DATABASE_URL not found in .env');
 
-  await prisma.user.create({
-    data: {
-      email: 'leo@example.com',
-      password: 'leo_password_1',
-      posts: {
-        create: [
-          { message: 'Holaaaa Soy Leo ' },
-          { message: 'aqui estoy probando los posts' },
-        ],
+  const { pool, adapter } = createPrismaAdapter(dbUrl);
+  const prisma = new PrismaClient({ adapter });
+
+  const DEFAULT_PASSWORD = 'password123';
+  const hashedPassword = await bcrypt.hash(DEFAULT_PASSWORD, 10);
+
+  try {
+    console.log('Starting database seeding...');
+
+    await prisma.post.deleteMany();
+    await prisma.user.deleteMany();
+
+
+    await prisma.user.create({
+      data: {
+        email: 'leo@example.com',
+        password: hashedPassword,
+        posts: {
+          create: [
+            { message: 'Holaaaa Soy Leo' },
+            { message: 'Aquí estoy probando los posts' },
+          ],
+        },
       },
-    },
-  });
+    });
 
-  await prisma.user.create({
-    data: {
-      email: 'richard@example.com',
-      password: 'richard_password_2',
-      posts: {
-        create: [{ message: 'Holaaa soy Richard!' }],
+    await prisma.user.create({
+      data: {
+        email: 'richard@example.com',
+        password: hashedPassword,
+        posts: {
+          create: [{ message: 'Holaaa soy Richard!' }],
+        },
       },
-    },
-  });
+    });
 
-  console.log('✅ Database seeded');
+    console.log('Database seeded successfully');
+  } catch (error) {
+    console.error('Error during seeding:', error);
+    process.exit(1);
+  } finally {
+    await prisma.$disconnect();
+    await pool.end();
+  }
 }
 
 main()
-  .catch((e) => {
-    console.error(e);
-    process.exit(1);
-  })
-  .finally(async () => {
-    await prisma.$disconnect();
-    await pool.end();
-  });
